Refactor:  funciones para excel y controllers
* refactor: materia repository, useCase saveMateria

* Refactor: Se elimino saveBatch del repository de materia y se puso como un caso de uso. Tambien se refactorizo UploadPlanEstudio y se ubico en su propia carpeta

* Refactor: se refactorizo el controller de materia con el uploadExcel

+++ backend/src/application/useCases/relacionImparte/DeleteRelacionImparte.ts
@@ -17,8 +17,8 @@ export class DeleteRelacionImparte {
     if (term.trim() === '') {
       throw new Error('El término es requerido para eliminar')
     }
-    if (nroSeccion <= 0) {
-      throw new Error('El número de sección es requerido para eliminar')
+    if (isNaN(nroSeccion) || nroSeccion <= 0) {
+      throw new Error('El número de sección debe ser un número entero válido y mayor a 0 para eliminar')
     }
 
     await this.repository.delete(cedulaP, codAsig, term, nroSeccion)
